+++ core/src/structs/trade.rs
@@ -6,7 +6,7 @@ use crate::structs::market::IsCommodity;
 #[derive(Debug)]
 pub struct Trade<C: IsCommodity, I: Clone> {
     pub commodity: C,
-    pub volume: f32,
+    pub volume: Decimal,
     pub price: Decimal,
     pub timestamp: DateTime<FixedOffset>,
     pub seller: I,
